Move JSONObjectParceler to utils
