+++ public/src/test/java/com/revenuecat/purchases/ParcelableTests.kt
@@ -3,7 +3,6 @@ package com.revenuecat.purchases
 import android.net.Uri
 import android.os.Parcel
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.revenuecat.purchases.parceler.JSONObjectParceler
 import com.revenuecat.purchases.parceler.JSONObjectParceler.write
 import com.revenuecat.purchases.utils.Responses
 import com.revenuecat.purchases.utils.testParcelization
